Remote thrown exception in Entity::commit()
This patch removes the exception that was being thrown in
Entity::commit(). We need to assume that the commit succeeds when no
rows are reportedly updated. This is because the DB engine won't update
a row with the same content.

This patch also removes the transaction since it probably wasn't needed
in the first place.