[Entity] Improve commit state management
This commit updates when the commit state is modified so that the entity
is in a more accurate state at the right time. Specifically, this allows
the entity's fetch, exist and create states to have the right values
before the postCommit() hook is called. Furthermore, we now allow users
to toggle field invalidation in the commit() call.