Implement reflection entity abstraction
This commit adds the first draft for a reflection-based entity class.
This class subclasses the Entity class and provides mechanisms for
loading the entity's structure from doc comments contained in the entity
class's class comment and property comments.

The Entity interface remained mostly the same. The implementation was
refactored slightly to account for new requirements. Mainly, the usage
of the 'updates' and 'fields' buckets in the Entity implementation
changed since the ReflectionEntity subclass does not need these buckets
to store state information.

The ReflectionEntityTrait trait is used for accessing private properties
of the Entity subclass. So users must both subclass ReflectionEntity and
import the ReflectionEntityTrait when creating custom entity types.