Entity: Fix a few issues
Fixes issue causing transaction to never commit when committing via a
recursive call. The recursive call is made when a test for entity
existence indicates it does not exist.

Fixes issue where commitSuccess() was not called before a postCommit()
in one of the cases. This was causing the entity to not be properly
invalidated when committing.