Remove dirty fields state management
This patch removes dirty fields state management from ReflectionEntity
subclass since it really wasn't needed in the first place. The updates
state from the Entity base class is more than sufficient for detecting
changes, so long as fields are appropriately touched by the Entity
subclass.

We added to Entity a mechanism for manipulating the updates state in
derived classes. This allows a derived class to use the updates state
instead of having to maintain its own dirty fields state.