EntityList: Improve commit update operation
This patch improves the commit update operation so that it uses prepared
statements with positional bound parameters. To account for potentially
different combinations of field updates, we generate as many prepared
statements as are necessary to independently update each entity.

This is a preferred implementation over using a single UPDATE statement
with many WHEN...CASE clauses since we can use positional parameters.
Filter variables must be indicated using positional parameters, so each
of the commit queries (i.e. insert, update and delete) must utilize
positional parameters as well.